requirements.txt
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3
